requirements.txt
opencv-python==4.2.0.32
tqdm==4.36.1
h5py==2.10.0
matplotlib==3.1.2
numpy==1.17.4
tensorboardX==1.9
torch==1.3.1
torchvision==0.4.2
pillow==7.1.1
scipy==1.4.1
tensorboard==1.14